Rename TUNE's three unchanged screens to their new section names
VeScreen -> AirflowScreen, TimingScreen -> SparkScreen, AfrScreen -> FuelScreen,
each with its .module.css. EcuLab.jsx's imports and mount conditions now key off
the new section ids (airflow/spark/fuel) while keeping the mutually-exclusive
conditional-render shape. tune-screens.test.jsx re-pointed to the new names with
no assertion changes.
